Refactor: Link editing (use unique IDs)
This one was a big deal. Previously, there was an array of selected
words stored in application state that would be used to create
new links. Now, we store the inProgress link in state and
mutate it based on user action. When the user finalizes the link,
we lookup up the link by ID and replace or insert.

This resulted in simpler react components and heaver reducers.
Reducers are usually easier to test, so I like this tradeoff for now.
RTK w/ Immer makes it easier to write readable reducers with
branching logic. This wasn't so much the case using switch
statements and pure immutability.