NVI is now a source and LEB + SBL GNT align to it.
This was in interesting test of the framework that I put in place
for the specification of which corpus aligns to which. I think it
remains confusing that we end up calling a translations a
'source', (I might prefer 'hub' to 'source') but as long as one
gets in the right headspace, it works.